♻️ refactor: Refactor Dockerfile installation process and add additional LDAP schemas
The commit includes changes to the Dockerfile, adding support for installing additional LDAP schemas and refactoring the installation process. It also modifies the configuration files for access control and password policies. Additionally, it introduces LDIF templates for groups and organization, as well as renaming the "users.ldif.tmpl" file to "people.ldif.tmpl". The docker-compose.yaml file is modified to update the container names and set a different port for phpldapadmin.

Please note that the commit message assumes the context and significance of the changes based on the provided diff.