✨ feat: Update Dockerfile and entrypoint.sh
- Updated Dockerfile to copy the 'migrations' directory instead of 'config'
- Updated entrypoint.sh to apply migrations instead of config
- Renamed files in 'context/config' directory to 'context/migrations'
- Updated docker-compose.yaml to include volume mapping for openldap

This commit modifies the Dockerfile and entrypoint.sh script to use the 'migrations' directory instead of the 'config' directory. It also renames the files in the 'context/config' directory to 'context/migrations' for consistency. Additionally, the docker-compose.yaml file is updated to include a volume mapping for the 'openldap' service.

Co-authored-by: Clay Risser <email@clayrisser.com>